Renaming file to match the name of the class
